+++ test/e2e/rte/local/local.go
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package rte
+package local
 
 import (
 	"encoding/json"
